Move testing wrapper into separate file.
